added delete alarm verification

+++ client/src/components/AlarmManager.tsx
@@ -13,6 +13,7 @@ import {
 const AlarmManager = () => {
   const [alarms, setAlarms] = useState<AlarmManagerProps[]>([]);
   const [editingAlarmId, setEditingAlarmId] = useState<number | null>(null);
+  const [confirmDeleteId, setConfirmDeleteId] = useState<number | null>(null);
 
   useEffect(() => {
     loadAlarms();
@@ -36,12 +37,22 @@ const AlarmManager = () => {
     }
   };
 
+  const handleConfirmDelete = (id: number) => {
+    setConfirmDeleteId(id);
+    // Optionally set a timeout to reset the confirmation state
+    setTimeout(() => setConfirmDeleteId(null), 5000); // 5 seconds to confirm
+  };
+
   const handleDeleteAlarm = async (id: number) => {
-    try {
-      await deleteAlarm(id);
-      setAlarms(alarms.filter((alarm) => alarm.id !== id));
-    } catch (error) {
-      console.error("Couldn't delete alarm:", error);
+    // If there's a confirmation id, only delete if it matches
+    if (confirmDeleteId === id) {
+      try {
+        await deleteAlarm(id);
+        setAlarms(alarms.filter((alarm) => alarm.id !== id));
+        setConfirmDeleteId(null); // Reset confirmation state after deletion
+      } catch (error) {
+        console.error("Couldn't delete alarm:", error);
+      }
     }
   };
 
@@ -104,23 +115,43 @@ const AlarmManager = () => {
                         <path d='M21.731 2.269a2.625 2.625 0 0 0-3.712 0l-1.157 1.157 3.712 3.712 1.157-1.157a2.625 2.625 0 0 0 0-3.712ZM19.513 8.199l-3.712-3.712-12.15 12.15a5.25 5.25 0 0 0-1.32 2.214l-.8 2.685a.75.75 0 0 0 .933.933l2.685-.8a5.25 5.25 0 0 0 2.214-1.32L19.513 8.2Z' />
                       </svg>
                     </button>
-                    <button
-                      onClick={() => handleDeleteAlarm(alarm.id)}
-                      className='p-2 rounded-full text-red-500 hover:text-red-700 focus:outline-none focus:ring focus:ring-red-300 ml-3'
-                    >
-                      <svg
-                        xmlns='http://www.w3.org/2000/svg'
-                        viewBox='0 0 24 24'
-                        fill='currentColor'
-                        className='w-6 h-6'
+                    {confirmDeleteId === alarm.id ? (
+                      <button
+                        onClick={() => handleDeleteAlarm(alarm.id)}
+                        className='p-2 rounded-full text-green-500 hover:text-green-700 focus:outline-none  ml-3'
                       >
-                        <path
-                          fill-rule='evenodd'
-                          d='M16.5 4.478v.227a48.816 48.816 0 0 1 3.878.512.75.75 0 1 1-.256 1.478l-.209-.035-1.005 13.07a3 3 0 0 1-2.991 2.77H8.084a3 3 0 0 1-2.991-2.77L4.087 6.66l-.209.035a.75.75 0 0 1-.256-1.478A48.567 48.567 0 0 1 7.5 4.705v-.227c0-1.564 1.213-2.9 2.816-2.951a52.662 52.662 0 0 1 3.369 0c1.603.051 2.815 1.387 2.815 2.951Zm-6.136-1.452a51.196 51.196 0 0 1 3.273 0C14.39 3.05 15 3.684 15 4.478v.113a49.488 49.488 0 0 0-6 0v-.113c0-.794.609-1.428 1.364-1.452Zm-.355 5.945a.75.75 0 1 0-1.5.058l.347 9a.75.75 0 1 0 1.499-.058l-.346-9Zm5.48.058a.75.75 0 1 0-1.498-.058l-.347 9a.75.75 0 0 0 1.5.058l.345-9Z'
-                          clip-rule='evenodd'
-                        />
-                      </svg>
-                    </button>
+                        <svg
+                          xmlns='http://www.w3.org/2000/svg'
+                          viewBox='0 0 24 24'
+                          fill='currentColor'
+                          className='w-6 h-6'
+                        >
+                          <path
+                            fillRule='evenodd'
+                            d='M19.916 4.626a.75.75 0 0 1 .208 1.04l-9 13.5a.75.75 0 0 1-1.154.114l-6-6a.75.75 0 0 1 1.06-1.06l5.353 5.353 8.493-12.74a.75.75 0 0 1 1.04-.207Z'
+                            clipRule='evenodd'
+                          />
+                        </svg>
+                      </button>
+                    ) : (
+                      <button
+                        onClick={() => handleConfirmDelete(alarm.id)}
+                        className='p-2 rounded-full text-red-500 hover:text-red-700 focus:outline-none ml-3'
+                      >
+                        <svg
+                          xmlns='http://www.w3.org/2000/svg'
+                          viewBox='0 0 24 24'
+                          fill='currentColor'
+                          className='w-6 h-6'
+                        >
+                          <path
+                            fillRule='evenodd'
+                            d='M16.5 4.478v.227a48.816 48.816 0 0 1 3.878.512.75.75 0 1 1-.256 1.478l-.209-.035-1.005 13.07a3 3 0 0 1-2.991 2.77H8.084a3 3 0 0 1-2.991-2.77L4.087 6.66l-.209.035a.75.75 0 0 1-.256-1.478A48.567 48.567 0 0 1 7.5 4.705v-.227c0-1.564 1.213-2.9 2.816-2.951a52.662 52.662 0 0 1 3.369 0c1.603.051 2.815 1.387 2.815 2.951Zm-6.136-1.452a51.196 51.196 0 0 1 3.273 0C14.39 3.05 15 3.684 15 4.478v.113a49.488 49.488 0 0 0-6 0v-.113c0-.794.609-1.428 1.364-1.452Zm-.355 5.945a.75.75 0 1 0-1.5.058l.347 9a.75.75 0 1 0 1.499-.058l-.346-9Zm5.48.058a.75.75 0 1 0-1.498-.058l-.347 9a.75.75 0 0 0 1.5.058l.345-9Z'
+                            clipRule='evenodd'
+                          />
+                        </svg>
+                      </button>
+                    )}
                   </div>
                 </div>
               )}
